AUTORIZACAO - Validando no controller

+++ app/Http/Controllers/ProjectController.php
@@ -36,7 +36,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id'=>Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -58,6 +58,9 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if ($this->checkProjectOwner($id) == false){
+            return ['error' => 'Access forbidden'];
+        }
         return $this->repository->find($id);
     }
 
@@ -70,6 +73,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if ($this->checkProjectOwner($id) == false){
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->update($request->all(),$id);
     }
 
@@ -81,6 +87,15 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if ($this->checkProjectOwner($id) == false){
+            return ['error' => 'Access forbidden'];
+        }
         $this->repository->delete($id);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectId, $userId);
     }
 }
